Added a random move generator
Used the rand crate to implement a random move generator. Mostly without bugs. The only problem is when you play against it as Circles. It moves first but the screen only updates after you move. So you can't play around it's moves. Also when it wins it prints two boards with no breaks in between them. Making it hard to read. This problem does not happen when you play as X.