The main game is finished.
The main game works without the engine and with it. There is no weirdness with the printing of the board nor any problems with the game logic. Now that the base program is done it's time to move onto a real engine. One that makes decisions and is not random.